Supported Genomes

Lists supported genomes

Usage

supportedGenomes()

Value

A data.frame containing supported genomes.

Details

Uses the ucscGenomes() function from the rtracklayer package to obtain a list of genomes available from the UCSC genome browser. The db column lists genomes as they are provided to the genome arguement of various functions.

The final column, headed AvailableGeneIDs lists the gene ID formats for which there is a local copy of the length information avaible for the genome listed in the geneLenDataBase package.

Author

Matthew D. Young myoung@wehi.edu.au

Examples

supportedGenomes()
#>          db          species       date
#> 3   anoCar1           Lizard  Feb. 2007
#> 5   anoGam1       A. gambiae  Feb. 2003
#> 7   apiMel1     A. mellifera  July 2004
#> 8   apiMel2     A. mellifera  Jan. 2005
#> 9   aplCal1         Sea hare Sept. 2008
#> 14  bosTau2              Cow  Mar. 2005
#> 15  bosTau3              Cow  Aug. 2006
#> 16  bosTau4              Cow  Oct. 2007
#> 21  braFlo1         Lancelet  Mar. 2006
#> 22  caeJap1      C. japonica  Mar. 2008
#> 23   caePb1      C. brenneri  Jan. 2007
#> 24   caePb2      C. brenneri  Feb. 2008
#> 25  caeRem2       C. remanei  Mar. 2006
#> 26  caeRem3       C. remanei   May 2007
#> 27  calJac1         Marmoset  June 2007
#> 31  canFam1              Dog  July 2004
#> 32  canFam2              Dog   May 2005
#> 37  cavPor3       Guinea pig  Feb. 2008
#> 38      cb1      C. briggsae  July 2002
#> 39      cb3      C. briggsae  Jan. 2007
#> 42      ce2       C. elegans  Mar. 2004
#> 43      ce4       C. elegans  Jan. 2007
#> 44      ce6       C. elegans   May 2008
#> 49      ci1  C. intestinalis  Dec. 2002
#> 50      ci2  C. intestinalis  Mar. 2005
#> 57  danRer3        Zebrafish   May 2005
#> 58  danRer4        Zebrafish  Mar. 2006
#> 59  danRer5        Zebrafish  July 2007
#> 60  danRer6        Zebrafish  Dec. 2008
#> 64      dm1  D. melanogaster  Jan. 2003
#> 65      dm2  D. melanogaster  Apr. 2004
#> 66      dm3  D. melanogaster  Apr. 2006
#> 68      dp2 D. pseudoobscura  Aug. 2003
#> 69      dp3 D. pseudoobscura  Nov. 2004
#> 70  droAna1     D. ananassae  July 2004
#> 71  droAna2     D. ananassae  Aug. 2005
#> 72  droEre1        D. erecta  Aug. 2005
#> 73  droGri1     D. grimshawi  Aug. 2005
#> 74  droMoj1    D. mojavensis  Aug. 2004
#> 75  droMoj2    D. mojavensis  Aug. 2005
#> 76  droPer1    D. persimilis  Oct. 2005
#> 77  droSec1     D. sechellia  Oct. 2005
#> 78  droSim1      D. simulans  Apr. 2005
#> 79  droVir1       D. virilis  July 2004
#> 80  droVir2       D. virilis  Aug. 2005
#> 81  droYak1        D. yakuba  Apr. 2004
#> 82  droYak2        D. yakuba  Nov. 2005
#> 87  equCab1            Horse  Jan. 2007
#> 88  equCab2            Horse  Sep. 2007
#> 92  felCat3              Cat  Mar. 2006
#> 97      fr1             Fugu  Aug. 2002
#> 98      fr2             Fugu  Oct. 2004
#> 101 galGal2          Chicken  Feb. 2004
#> 102 galGal3          Chicken   May 2006
#> 107 gasAcu1      Stickleback  Feb. 2006
#> 115    hg16            Human  July 2003
#> 116    hg17            Human   May 2004
#> 117    hg18            Human  Mar. 2006
#> 118    hg19            Human  Feb. 2009
#> 122 loxAfr3         Elephant  Jul. 2009
#> 133     mm7            Mouse  Aug. 2005
#> 134     mm8            Mouse  Feb. 2006
#> 135     mm9            Mouse  July 2007
#> 136 monDom1          Opossum  Oct. 2004
#> 137 monDom4          Opossum  Jan. 2006
#> 138 monDom5          Opossum  Oct. 2006
#> 150 ornAna1         Platypus  Mar. 2007
#> 153 oryLat2           Medaka  Oct. 2005
#> 161 panTro1            Chimp  Nov. 2003
#> 162 panTro2            Chimp  Mar. 2006
#> 170 petMar1          Lamprey  Mar. 2007
#> 173 ponAbe2        Orangutan  July 2007
#> 175 priPac1     P. pacificus  Feb. 2007
#> 179 rheMac2           Rhesus  Jan. 2006
#> 183     rn3              Rat  June 2003
#> 184     rn4              Rat  Nov. 2004
#> 188 sacCer1    S. cerevisiae  Oct. 2003
#> 189 sacCer2    S. cerevisiae  June 2008
#> 196 strPur1    S. purpuratus  Apr. 2005
#> 197 strPur2    S. purpuratus  Sep. 2006
#> 201 taeGut1      Zebra finch  Jul. 2008
#> 205 tetNig1        Tetraodon  Feb. 2004
#> 206 tetNig2        Tetraodon  Mar. 2007
#> 215 xenTro1    X. tropicalis  Oct. 2004
#> 217 xenTro2    X. tropicalis  Aug. 2005
#>                                                                                    name
#> 3                                                          Broad Institute AnoCar (1.0)
#> 5                                                                          IAGEC v.MOZ2
#> 7                                                                  Baylor HGSC Amel_1.2
#> 8                                                                  Baylor HGSC Amel_2.0
#> 9                                                                    Broad v. Aplcal2.0
#> 14                                                                      Baylor Btau_2.0
#> 15                                                                   Baylor Release 3.1
#> 16                                                                   Baylor Release 4.0
#> 21                                                               JGI v.1.0 (March 2006)
#> 22                        Washington University School of Medicine GSC C. remanei 3.0.2
#> 23            Washington University School of Medicine GSC Caenorhabditis n. sp. PB2801
#> 24  Washington University School of Medicine GSC Caenorhabditis brenneri (PB2801) 6.0.1
#> 25                          Washington University School of Medicine GSC C. remanei 1.0
#> 26                       Washington University School of Medicine GSC C. remanei 15.0.1
#> 27                                                                          WUSTL 2.0.2
#> 31                                                               Broad Institute v. 1.0
#> 32                                                               Broad Institute v. 2.0
#> 37                                                              Broad Institute cavPor3
#> 38                                                                WormBase v. cb25.agp8
#> 39                Washington University School of Medicine GSC and Sanger Institute cb3
#> 42                                                                    WormBase v. WS120
#> 43              Washington University School of Medicine GSC and Sanger Institute WS170
#> 44              Washington University School of Medicine GSC and Sanger Institute WS190
#> 49                                                                             JGI v1.0
#> 50                                                                             JGI v2.0
#> 57                                    Sanger Centre, Danio rerio Sequencing Project Zv5
#> 58                                    Sanger Centre, Danio rerio Sequencing Project Zv6
#> 59                                        Wellcome Trust Sanger Institute, Zv7 assembly
#> 60                       Wellcome Trust Sanger Institute, Zv8 assembly (CAAK00000000.5)
#> 64                                                                            BDGP v. 3
#> 65                                                              BDGP v. 4 / DHGP v. 3.2
#> 66                                                                       BDGP Release 5
#> 68                                                                 Baylor HGSC Freeze 1
#> 69                                                                 FlyBase Release 1.03
#> 70                                                                    TIGR 15 July 2004
#> 71                                                              Agencourt 1 August 2005
#> 72                                                                 Agencourt 1 Aug 2005
#> 73                                                                 Agencourt 1 Aug 2005
#> 74                                                                Agencourt 11 Aug 2004
#> 75                                                              Agencourt 1 August 2005
#> 76                                                                   Broad 28 Oct. 2005
#> 77                                                                   Broad 28 Oct. 2005
#> 78                                                                    WUSTL version 1.0
#> 79                                                               Agencourt 12 July 2004
#> 80                                                              Agencourt 1 August 2005
#> 81                                                                    WUSTL version 1.0
#> 82                                                                    WUSTL version 2.0
#> 87                                                        Broad Institute equCab1 (1.0)
#> 88                                                              Broad Institute EquCab2
#> 92                                                                      Broad Release 3
#> 97                                                                             JGI V3.0
#> 98                                                                             JGI V4.0
#> 101                              Chicken Genome Sequencing Consortium Feb. 2004 release
#> 102                               Chicken Genome Sequencing Consortium May 2006 release
#> 107                    Broad Institute gasAcu1 (1.0) (NCBI project 13579, AANH01000000)
#> 115                                                                       NCBI Build 34
#> 116                                                                       NCBI Build 35
#> 117                                                                     NCBI Build 36.1
#> 118             GRCh37 Genome Reference Consortium Human Reference 37 (GCA_000001405.1)
#> 122                          Broad Institute loxAfr3 (NCBI project 12569, AAGU03000000)
#> 133                                                                       NCBI Build 35
#> 134                                                                       NCBI Build 36
#> 135                                                                       NCBI Build 37
#> 136                                                            Broad Inst. Prelim Oct04
#> 137                                                           Broad Inst. monDom4 Jan06
#> 138                Broad Institute monDom5 (NCBI project 12561, accession AAFR03000000)
#> 150                                                                 WUSTL version 5.0.1
#> 153                       NIG and Tokyo Univ MEDAKA1 (NCBI project 16702, BAAF04000000)
#> 161                                             CGSC Build 1 Version 1 Arachne assembly
#> 162                                             Washington University Build 2 Version 1
#> 170                                                            WUSTL v.3.0 (March 2007)
#> 173                                                                           WUSTL 3.0
#> 175             Washington University School of Medicine GSC Pristionchus pacificus 5.0
#> 179                         Venter Inst./Baylor/WashU Merged assembly Jan 2006 v1 edit4
#> 183                                                                  Baylor HGSC v. 3.1
#> 184                                                                  Baylor HGSC v. 3.4
#> 188                                                 SGD sequence data as of 1 Oct. 2003
#> 189                                                                       SGD June 2008
#> 196                                                                              Baylor
#> 197                                                           Baylor release 3 Spur 2.1
#> 201                                                       Washington University taeGut1
#> 205                 Genoscope and Broad Institute V7 (NCBI project 12350, CAAE01000000)
#> 206                         Genoscope Tetraodon v8.0 (NCBI project 12350, CAAE01000000)
#> 215                                                                            JGI v3.0
#> 217                                                                            JGI v4.1
#>                                                                                                                      AvailableGeneIDs
#> 3                                                                                                         ensGene,genscan,xenoRefGene
#> 5                                                                                                              ensGene,geneid,genscan
#> 7                                                                                                                             genscan
#> 8                                                                                                              ensGene,geneid,genscan
#> 9                                                                                                                         xenoRefGene
#> 14                                                                                          geneSymbol,geneid,genscan,refGene,sgpGene
#> 15                                                                                  ensGene,geneSymbol,geneid,genscan,refGene,sgpGene
#> 16                                                                                       ensGene,geneSymbol,genscan,nscanGene,refGene
#> 21                                                                                                                        xenoRefGene
#> 22                                                                                                                        xenoRefGene
#> 23                                                                                                                        xenoRefGene
#> 24                                                                                                                        xenoRefGene
#> 25                                                                                                                        xenoRefGene
#> 26                                                                                                                        xenoRefGene
#> 27                                                                                                      genscan,nscanGene,xenoRefGene
#> 31                                                                           ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 32                                                                           ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 37                                                                                              ensGene,genscan,nscanGene,xenoRefGene
#> 38                                                                                                                        xenoRefGene
#> 39                                                                                                                        xenoRefGene
#> 42                                                                                                          geneSymbol,geneid,refGene
#> 43                                                                                                     geneSymbol,refGene,xenoRefGene
#> 44                                                                                             ensGene,geneSymbol,refGene,xenoRefGene
#> 49                                                                                                     geneSymbol,refGene,xenoRefGene
#> 50                                                                                             ensGene,geneSymbol,refGene,xenoRefGene
#> 57                                                                                                         ensGene,geneSymbol,refGene
#> 58                                                                                       ensGene,geneSymbol,genscan,nscanGene,refGene
#> 59                                                                                 ensGene,geneSymbol,refGene,vegaGene,vegaPseudoGene
#> 60                                                                                             ensGene,geneSymbol,refGene,xenoRefGene
#> 64                                                                                                         geneSymbol,genscan,refGene
#> 65                                                                                        geneSymbol,geneid,genscan,nscanGene,refGene
#> 66                                                                                                   geneSymbol,nscanPasaGene,refGene
#> 68                                                                                                                genscan,xenoRefGene
#> 69                                                                                                         geneid,genscan,xenoRefGene
#> 70                                                                                                         geneid,genscan,xenoRefGene
#> 71                                                                                                                genscan,xenoRefGene
#> 72                                                                                                                genscan,xenoRefGene
#> 73                                                                                                                genscan,xenoRefGene
#> 74                                                                                                         geneid,genscan,xenoRefGene
#> 75                                                                                                                genscan,xenoRefGene
#> 76                                                                                                                genscan,xenoRefGene
#> 77                                                                                                                genscan,xenoRefGene
#> 78                                                                                                         geneid,genscan,xenoRefGene
#> 79                                                                                                         geneid,genscan,xenoRefGene
#> 80                                                                                                                genscan,xenoRefGene
#> 81                                                                                                         geneid,genscan,xenoRefGene
#> 82                                                                                                                genscan,xenoRefGene
#> 87                                                                                        geneSymbol,geneid,nscanGene,refGene,sgpGene
#> 88                                                                                   ensGene,geneSymbol,nscanGene,refGene,xenoRefGene
#> 92                                                            ensGene,geneSymbol,geneid,genscan,nscanGene,refGene,sgpGene,xenoRefGene
#> 97                                                                                                                    ensGene,genscan
#> 98                                                                                                                            ensGene
#> 101                                                                                 ensGene,geneSymbol,geneid,genscan,refGene,sgpGene
#> 102                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 107                                                                                                                 ensGene,nscanGene
#> 115                                                      acembly,ensGene,exoniphy,geneSymbol,geneid,genscan,knownGene,refGene,sgpGene
#> 116 acembly,acescan,ccdsGene,ensGene,exoniphy,geneSymbol,geneid,genscan,knownGene,refGene,sgpGene,vegaGene,vegaPseudoGene,xenoRefGene
#> 117   acembly,acescan,ccdsGene,ensGene,exoniphy,geneSymbol,geneid,genscan,knownGene,knownGeneOld3,refGene,sgpGene,sibGene,xenoRefGene
#> 118                                                      ccdsGene,ensGene,exoniphy,geneSymbol,knownGene,nscanGene,refGene,xenoRefGene
#> 122                                                                                                                       xenoRefGene
#> 133                                                           ensGene,geneSymbol,geneid,genscan,knownGene,refGene,sgpGene,xenoRefGene
#> 134                                ccdsGene,ensGene,geneSymbol,geneid,genscan,knownGene,nscanGene,refGene,sgpGene,sibGene,xenoRefGene
#> 135                       acembly,ccdsGene,ensGene,exoniphy,geneSymbol,geneid,genscan,knownGene,nscanGene,refGene,sgpGene,xenoRefGene
#> 136                                                                                                                           genscan
#> 137                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 138                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 150                                                                                            ensGene,geneSymbol,refGene,xenoRefGene
#> 153                                                                                            ensGene,geneSymbol,refGene,xenoRefGene
#> 161                                                                                                ensGene,geneid,genscan,xenoRefGene
#> 162                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 170                                                                                                                       xenoRefGene
#> 173                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 175                                                                                                                       xenoRefGene
#> 179                                                                   ensGene,geneSymbol,geneid,nscanGene,refGene,sgpGene,xenoRefGene
#> 183                                                 ensGene,geneSymbol,geneid,genscan,knownGene,nscanGene,refGene,sgpGene,xenoRefGene
#> 184                                                 ensGene,geneSymbol,geneid,genscan,knownGene,nscanGene,refGene,sgpGene,xenoRefGene
#> 188                                                                                                                           ensGene
#> 189                                                                                                                           ensGene
#> 196                                                                                            geneSymbol,genscan,refGene,xenoRefGene
#> 197                                                                                            geneSymbol,genscan,refGene,xenoRefGene
#> 201                                                                          ensGene,geneSymbol,genscan,nscanGene,refGene,xenoRefGene
#> 205                                                                                                  ensGene,geneid,genscan,nscanGene
#> 206                                                                                                                           ensGene
#> 215                                                                                                                           genscan
#> 217                                                                                                ensGene,geneSymbol,genscan,refGene