Purge factors

Removes all factors from a variable in a sensible way.

Usage

unfactor(var)

Arguments

var

The variable from which you want the factors removed.

Value

The variable with all factors converted to characters or numbers (see details).

Details

As factors are their own type, to remove factors we must convert each level into another type. This is currently done using "typeless" behaviour: a factor is converted to a numeric vector if this can be done without inducing NAs, otherwise it is coerced using as.character. Currently supported types are: factor, data.frame and list.

Author

Matthew D. Young myoung@wehi.edu.au

Examples

# A named factor
x <- factor(sample(1:6, 100, replace=TRUE))
names(x) <- paste("Roll.No", 1:100, sep='.')
x
#>   Roll.No.1   Roll.No.2   Roll.No.3   Roll.No.4   Roll.No.5   Roll.No.6 
#>           5           4           5           6           4           6 
#>   Roll.No.7   Roll.No.8   Roll.No.9  Roll.No.10  Roll.No.11  Roll.No.12 
#>           1           5           5           5           2           6 
#>  Roll.No.13  Roll.No.14  Roll.No.15  Roll.No.16  Roll.No.17  Roll.No.18 
#>           3           6           3           5           5           4 
#>  Roll.No.19  Roll.No.20  Roll.No.21  Roll.No.22  Roll.No.23  Roll.No.24 
#>           5           2           6           4           2           3 
#>  Roll.No.25  Roll.No.26  Roll.No.27  Roll.No.28  Roll.No.29  Roll.No.30 
#>           1           6           3           5           6           1 
#>  Roll.No.31  Roll.No.32  Roll.No.33  Roll.No.34  Roll.No.35  Roll.No.36 
#>           6           2           5           2           1           4 
#>  Roll.No.37  Roll.No.38  Roll.No.39  Roll.No.40  Roll.No.41  Roll.No.42 
#>           6           1           6           2           3           6 
#>  Roll.No.43  Roll.No.44  Roll.No.45  Roll.No.46  Roll.No.47  Roll.No.48 
#>           1           1           2           5           1           4 
#>  Roll.No.49  Roll.No.50  Roll.No.51  Roll.No.52  Roll.No.53  Roll.No.54 
#>           3           4           4           3           2           3 
#>  Roll.No.55  Roll.No.56  Roll.No.57  Roll.No.58  Roll.No.59  Roll.No.60 
#>           2           2           3           6           4           2 
#>  Roll.No.61  Roll.No.62  Roll.No.63  Roll.No.64  Roll.No.65  Roll.No.66 
#>           1           6           6           3           1           2 
#>  Roll.No.67  Roll.No.68  Roll.No.69  Roll.No.70  Roll.No.71  Roll.No.72 
#>           1           4           1           4           6           1 
#>  Roll.No.73  Roll.No.74  Roll.No.75  Roll.No.76  Roll.No.77  Roll.No.78 
#>           2           2           6           4           2           4 
#>  Roll.No.79  Roll.No.80  Roll.No.81  Roll.No.82  Roll.No.83  Roll.No.84 
#>           6           6           5           3           4           1 
#>  Roll.No.85  Roll.No.86  Roll.No.87  Roll.No.88  Roll.No.89  Roll.No.90 
#>           5           1           5           6           1           1 
#>  Roll.No.91  Roll.No.92  Roll.No.93  Roll.No.94  Roll.No.95  Roll.No.96 
#>           5           2           5           3           1           4 
#>  Roll.No.97  Roll.No.98  Roll.No.99 Roll.No.100 
#>           2           1           3           1 
#> Levels: 1 2 3 4 5 6
unfactor(x)
#>   Roll.No.1   Roll.No.2   Roll.No.3   Roll.No.4   Roll.No.5   Roll.No.6 
#>           5           4           5           6           4           6 
#>   Roll.No.7   Roll.No.8   Roll.No.9  Roll.No.10  Roll.No.11  Roll.No.12 
#>           1           5           5           5           2           6 
#>  Roll.No.13  Roll.No.14  Roll.No.15  Roll.No.16  Roll.No.17  Roll.No.18 
#>           3           6           3           5           5           4 
#>  Roll.No.19  Roll.No.20  Roll.No.21  Roll.No.22  Roll.No.23  Roll.No.24 
#>           5           2           6           4           2           3 
#>  Roll.No.25  Roll.No.26  Roll.No.27  Roll.No.28  Roll.No.29  Roll.No.30 
#>           1           6           3           5           6           1 
#>  Roll.No.31  Roll.No.32  Roll.No.33  Roll.No.34  Roll.No.35  Roll.No.36 
#>           6           2           5           2           1           4 
#>  Roll.No.37  Roll.No.38  Roll.No.39  Roll.No.40  Roll.No.41  Roll.No.42 
#>           6           1           6           2           3           6 
#>  Roll.No.43  Roll.No.44  Roll.No.45  Roll.No.46  Roll.No.47  Roll.No.48 
#>           1           1           2           5           1           4 
#>  Roll.No.49  Roll.No.50  Roll.No.51  Roll.No.52  Roll.No.53  Roll.No.54 
#>           3           4           4           3           2           3 
#>  Roll.No.55  Roll.No.56  Roll.No.57  Roll.No.58  Roll.No.59  Roll.No.60 
#>           2           2           3           6           4           2 
#>  Roll.No.61  Roll.No.62  Roll.No.63  Roll.No.64  Roll.No.65  Roll.No.66 
#>           1           6           6           3           1           2 
#>  Roll.No.67  Roll.No.68  Roll.No.69  Roll.No.70  Roll.No.71  Roll.No.72 
#>           1           4           1           4           6           1 
#>  Roll.No.73  Roll.No.74  Roll.No.75  Roll.No.76  Roll.No.77  Roll.No.78 
#>           2           2           6           4           2           4 
#>  Roll.No.79  Roll.No.80  Roll.No.81  Roll.No.82  Roll.No.83  Roll.No.84 
#>           6           6           5           3           4           1 
#>  Roll.No.85  Roll.No.86  Roll.No.87  Roll.No.88  Roll.No.89  Roll.No.90 
#>           5           1           5           6           1           1 
#>  Roll.No.91  Roll.No.92  Roll.No.93  Roll.No.94  Roll.No.95  Roll.No.96 
#>           5           2           5           3           1           4 
#>  Roll.No.97  Roll.No.98  Roll.No.99 Roll.No.100 
#>           2           1           3           1 

# A data.frame
x <- data.frame(player <- c("Alice", "Bob", "Mary", "Fred"), 
                score <- factor(c(9, 7, 8, 9)), stringsAsFactors=TRUE)
x$player
#> [1] Alice Bob   Mary  Fred 
#> Levels: Alice Bob Fred Mary
x$score
#> [1] 9 7 8 9
#> Levels: 7 8 9
y <- unfactor(x)
y$player
#> [1] "Alice" "Bob"   "Mary"  "Fred" 
y$score
#> [1] 9 7 8 9